docs_site/requirements.txt
flask>=2.3.0
ollama>=0.1.0
groq>=0.4.0
python-dotenv>=1.0.0
markdown>=3.5.0
numpy>=1.24.0
